chore(repo): split e2e-workspace to two applications
