+++ e2e/workspace-core/project.json
@@ -1,6 +1,6 @@
 {
-  "root": "e2e/workspace",
-  "sourceRoot": "e2e/workspace",
+  "root": "e2e/workspace-core",
+  "sourceRoot": "e2e/workspace-core",
   "projectType": "application",
   "targets": {
     "e2e": {
@@ -14,7 +14,7 @@
             "command": "yarn e2e-build-package-publish"
           },
           {
-            "command": "nx run-e2e-tests e2e-workspace"
+            "command": "nx run-e2e-tests e2e-workspace-core"
           }
         ],
         "parallel": false
@@ -23,12 +23,12 @@
     "run-e2e-tests": {
       "executor": "@nrwl/jest:jest",
       "options": {
-        "jestConfig": "e2e/workspace/jest.config.js",
+        "jestConfig": "e2e/workspace-core/jest.config.js",
         "passWithNoTests": true,
         "runInBand": true
       },
-      "outputs": ["coverage/e2e/workspace"]
+      "outputs": ["coverage/e2e/workspace-core"]
     }
   },
-  "implicitDependencies": ["create-nx-workspace", "angular", "react"]
+  "implicitDependencies": []
 }
