+++ e2e/workspace-core/src/plugins.test.ts
