+++ e2e/workspace-core/src/workspace-lib.test.ts
