+++ e2e/workspace-core/tsconfig.json
