+++ e2e/workspace-core/tsconfig.spec.json
