+++ e2e/workspace-integrations/src/create-nx-workspace.test.ts
