+++ e2e/workspace-integrations/src/workspace-aux-commands.test.ts
@@ -7,13 +7,11 @@ import {
   readFile,
   readJson,
   removeProject,
-  renameFile,
   runCLI,
   runCLIAsync,
   tmpProjPath,
   uniq,
   updateFile,
-  workspaceConfigName,
 } from '@nrwl/e2e/utils';
 import type { NxJsonConfiguration } from '@nrwl/devkit';
 import { classify } from '@nrwl/workspace/src/utils/strings';
@@ -22,97 +20,6 @@ let proj: string;
 
 beforeAll(() => {
   proj = newProject();
-});
-
-describe('lint', () => {
-  it('lint should ensure module boundaries', () => {
-    const myapp = uniq('myapp');
-    const myapp2 = uniq('myapp2');
-    const mylib = uniq('mylib');
-    const lazylib = uniq('lazylib');
-    const invalidtaglib = uniq('invalidtaglib');
-    const validtaglib = uniq('validtaglib');
-
-    runCLI(`generate @nrwl/angular:app ${myapp} --tags=validtag`);
-    runCLI(`generate @nrwl/angular:app ${myapp2}`);
-    runCLI(`generate @nrwl/angular:lib ${mylib}`);
-    runCLI(`generate @nrwl/angular:lib ${lazylib}`);
-    runCLI(`generate @nrwl/angular:lib ${invalidtaglib} --tags=invalidtag`);
-    runCLI(`generate @nrwl/angular:lib ${validtaglib} --tags=validtag`);
-
-    const eslint = readJson('.eslintrc.json');
-    eslint.overrides[0].rules[
-      '@nrwl/nx/enforce-module-boundaries'
-    ][1].depConstraints = [
-      { sourceTag: 'validtag', onlyDependOnLibsWithTags: ['validtag'] },
-      ...eslint.overrides[0].rules['@nrwl/nx/enforce-module-boundaries'][1]
-        .depConstraints,
-    ];
-    updateFile('.eslintrc.json', JSON.stringify(eslint, null, 2));
-
-    const tsConfig = readJson('tsconfig.base.json');
-
-    /**
-     * apps do not add themselves to the tsconfig file.
-     *
-     * Let's add it so that we can trigger the lint failure
-     */
-    tsConfig.compilerOptions.paths[`@${proj}/${myapp2}`] = [
-      `apps/${myapp2}/src/main.ts`,
-    ];
-
-    tsConfig.compilerOptions.paths[`@secondScope/${lazylib}`] =
-      tsConfig.compilerOptions.paths[`@${proj}/${lazylib}`];
-    delete tsConfig.compilerOptions.paths[`@${proj}/${lazylib}`];
-    updateFile('tsconfig.base.json', JSON.stringify(tsConfig, null, 2));
-
-    updateFile(
-      `apps/${myapp}/src/main.ts`,
-      `
-      import '../../../libs/${mylib}';
-      import '@secondScope/${lazylib}';
-      import '@${proj}/${myapp2}';
-      import '@${proj}/${invalidtaglib}';
-      import '@${proj}/${validtaglib}';
-
-      const s = {loadChildren: '@${proj}/${lazylib}'};
-    `
-    );
-
-    const out = runCLI(`lint ${myapp}`, { silenceError: true });
-    expect(out).toContain(
-      'Libraries cannot be imported by a relative or absolute path, and must begin with a npm scope'
-    );
-    // expect(out).toContain('Imports of lazy-loaded libraries are forbidden');
-    expect(out).toContain('Imports of apps are forbidden');
-    expect(out).toContain(
-      'A project tagged with "validtag" can only depend on libs tagged with "validtag"'
-    );
-  }, 1000000);
-
-  describe('nx workspace-lint', () => {
-    it('should identify issues with the workspace', () => {
-      const appBefore = uniq('before');
-      const appAfter = uniq('after');
-
-      runCLI(`generate @nrwl/angular:app ${appBefore}`);
-      renameFile(`apps/${appBefore}`, `apps/${appAfter}`);
-
-      const stdout = runCLI('workspace-lint', { silenceError: true });
-      expect(stdout).toContain(
-        `- Cannot find project '${appBefore}' in 'apps/${appBefore}'`
-      );
-      expect(stdout).toContain(
-        'The following file(s) do not belong to any projects:'
-      );
-      expect(stdout).toContain(`- apps/${appAfter}/jest.config.js`);
-      expect(stdout).toContain(`- apps/${appAfter}/src/app/app.component.css`);
-      expect(stdout).toContain(`- apps/${appAfter}/src/app/app.component.html`);
-      expect(stdout).toContain(
-        `- apps/${appAfter}/src/app/app.component.spec.ts`
-      );
-    });
-  });
 });
 
 describe('format', () => {
@@ -244,159 +151,6 @@ describe('format', () => {
       );
     }
   }, 90000);
-});
-
-describe('workspace-generator', () => {
-  let custom: string;
-  let failing: string;
-
-  beforeEach(() => {
-    custom = uniq('custom');
-    failing = uniq('custom-failing');
-    runCLI(`g workspace-generator ${custom} --no-interactive`);
-    runCLI(`g workspace-generator ${failing} --no-interactive`);
-
-    checkFilesExist(
-      `tools/generators/${custom}/index.ts`,
-      `tools/generators/${custom}/schema.json`
-    );
-    checkFilesExist(
-      `tools/generators/${failing}/index.ts`,
-      `tools/generators/${failing}/schema.json`
-    );
-  });
-
-  it('should compile only generator files with dependencies', () => {
-    const workspace = uniq('workspace');
-
-    updateFile(
-      'tools/utils/utils.ts',
-      `
-        export const noop = () => {}
-        `
-    );
-    updateFile(
-      'tools/utils/logger.ts',
-      `
-        export const log = (...args: any[]) => console.log(...args)
-        `
-    );
-    updateFile(
-      `tools/generators/utils.ts`,
-      `
-        export const noop = ()=>{}
-        `
-    );
-    updateFile(`tools/generators/${custom}/index.ts`, (content) => {
-      return `
-          import { log } from '../../utils/logger'; \n
-          ${content}
-        `;
-    });
-
-    runCLI(`workspace-generator ${custom} ${workspace} --no-interactive -d`);
-
-    expect(() =>
-      checkFilesExist(
-        `dist/out-tsc/tools/generators/${custom}/index.js`,
-        `dist/out-tsc/tools/generators/utils.js`,
-        `dist/out-tsc/tools/utils/logger.js`
-      )
-    ).not.toThrow();
-    expect(() =>
-      checkFilesExist(`dist/out-tsc/tools/utils/utils.js`)
-    ).toThrow();
-  });
-
-  it('should support workspace-specific generators', async () => {
-    const json = readJson(`tools/generators/${custom}/schema.json`);
-    json.properties['directory'] = {
-      type: 'string',
-      description: 'lib directory',
-    };
-    json.properties['skipTsConfig'] = {
-      type: 'boolean',
-      description: 'skip changes to tsconfig',
-    };
-    updateFile(`tools/generators/${custom}/schema.json`, JSON.stringify(json));
-
-    const indexFile = readFile(`tools/generators/${custom}/index.ts`);
-    updateFile(
-      `tools/generators/${custom}/index.ts`,
-      indexFile.replace(
-        'name: schema.name',
-        'name: schema.name, directory: schema.directory, skipTsConfig: schema.skipTsConfig'
-      )
-    );
-
-    const workspace = uniq('workspace');
-    const dryRunOutput = runCLI(
-      `workspace-generator ${custom} ${workspace} --no-interactive --directory=dir --skipTsConfig=true -d`
-    );
-    expect(exists(`libs/dir/${workspace}/src/index.ts`)).toEqual(false);
-    expect(dryRunOutput).toContain(`UPDATE ${workspaceConfigName()}`);
-    expect(dryRunOutput).toContain('UPDATE nx.json');
-
-    const output = runCLI(
-      `workspace-generator ${custom} ${workspace} --no-interactive --directory=dir`
-    );
-    checkFilesExist(`libs/dir/${workspace}/src/index.ts`);
-    expect(output).toContain(`UPDATE ${workspaceConfigName()}`);
-    expect(output).toContain('UPDATE nx.json');
-
-    const jsonFailing = readJson(`tools/generators/${failing}/schema.json`);
-    jsonFailing.properties = {};
-    jsonFailing.required = [];
-    updateFile(
-      `tools/generators/${failing}/schema.json`,
-      JSON.stringify(jsonFailing)
-    );
-
-    updateFile(
-      `tools/generators/${failing}/index.ts`,
-      `
-          export default function() {
-            throw new Error();
-          }
-        `
-    );
-
-    try {
-      await runCLI(`workspace-generator ${failing} --no-interactive`);
-      fail(`Should exit 1 for a workspace-generator that throws an error`);
-    } catch (e) {}
-
-    const listOutput = runCLI('workspace-generator --list-generators');
-    expect(listOutput).toContain(custom);
-    expect(listOutput).toContain(failing);
-  }, 1000000);
-
-  it('should support angular devkit schematics', () => {
-    const angularDevkitSchematic = uniq('angular-devkit-schematic');
-    runCLI(`g workspace-generator ${angularDevkitSchematic} --no-interactive`);
-
-    const json = readJson(
-      `tools/generators/${angularDevkitSchematic}/schema.json`
-    );
-    json.properties = {};
-    json.required = [];
-    delete json.cli;
-    updateFile(
-      `tools/generators/${angularDevkitSchematic}/schema.json`,
-      JSON.stringify(json)
-    );
-
-    updateFile(
-      `tools/generators/${angularDevkitSchematic}/index.ts`,
-      `
-          export default function() {
-            return (tree) => tree;
-          }
-        `
-    );
-
-    runCLI(`workspace-generator ${angularDevkitSchematic} --no-interactive`);
-  });
 });
 
 describe('dep-graph', () => {
