+++ workspace.json
@@ -37,7 +37,8 @@
     "e2e-storybook": "e2e/storybook",
     "e2e-utils": "e2e/utils",
     "e2e-web": "e2e/web",
-    "e2e-workspace": "e2e/workspace",
+    "e2e-workspace-core": "e2e/workspace-core",
+    "e2e-workspace-integrations": "e2e/workspace-integrations",
     "eslint-plugin-nx": "packages/eslint-plugin-nx",
     "express": "packages/express",
     "gatsby": "packages/gatsby",
